requirements.txt
streamlit==1.32.0
torch==2.1.0
torchvision==0.16.0
timm==0.9.12
Pillow==9.5.0
opencv-python==4.9.0.80
numpy==1.24.4
grad-cam==1.4.8
shap==0.42.1
python-dotenv==1.0.1
requests==2.31.0
pandas==1.5.3
scikit-learn==1.3.2
matplotlib==3.7.1
